% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{The log-transformed summation of exponential values}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{A vector of value x increasing by 1 unit}
}
\value{
The log-transformed summation of the input vector
}
\description{
The log-transformed summation of exponential values
}
\examples{
log_summed_exps(c(1,2,3,4,5))
log_summed_exps(seq(1:5))
log_summed_exps(1:2000)

}
